payment APIを一つのファイルにまとめる
